requirements.txt
python-telegram-bot
yt-dlp
flask
requests